Largest architecture change since the beginning of the project: All the article interpretation code has been moved to /interpreter/.  All article interpretation is now done by Node.findAll / Node.findFirst - no more weird combination of SAX parsers and DOM crawling.  This makes the code a LOT easier to test and maintain!!
As part of this change, I rewrote how Node.findAll works - It now supports attribute selectors, so you can find meta tags with specific names or itemprops, or links with href attributes, or elements with any combination of classes and IDs.

Also, I switched Article to have paragraphs rather than article_body.  We already had the paragraphs, so it doesn't make sense to lose that information when putting it to storage.  This mean that I added support for repeated fields to our .data package - though I still need to work on having repeated fields be serializeable.

Lots of tests!!!