Fix Facebook share normalization (actually not sure if/how it was broken in retrospect... It may have just been that we were finding sites before they had social interaction yet, so everyone was having 0 interactions, so it ceased to be a signal).
So, realizing that, this change introduces a Scheduler task that will grab us up-to-date social engagements regularly during an article's infancy.  It's designed to run once an hour... It will crawl newer articles the most, until an article is 2.5 days old, then it stops (so we can focus on other new articles).